# ENH: Allow passing dictionaries of PhaseRecord objects to equilibrium and calculate