--- a/pycalphad/codegen/callables.py
+++ b/pycalphad/codegen/callables.py
@@ -146,7 +146,7 @@ def build_callables(dbf, comps, phases, models, parameter_symbols=None,
     return {output: _callables}
 
 
-def build_phase_records(dbf, comps, phases, conds, models, output='GM',
+def build_phase_records(dbf, comps, phases, state_variables, models, output='GM',
                         callables=None, parameters=None, verbose=False,
                         build_gradients=True, build_hessians=True
                         ):
@@ -161,10 +161,10 @@ def build_phase_records(dbf, comps, phases, conds, models, output='GM',
         List of active pure elements or species.
     phases : list
         List of phase names
-    conds : dict or None
-        Conditions for calculation
-    models : dict
-        Dictionary of {'phase_name': Model()}
+    state_variables : Iterable[v.StateVariable]
+        State variables used to produce the generated functions.
+    models : Mapping[str, Model]
+        Mapping of phase names to model instances
     parameters : dict, optional
         Maps SymPy Symbol to numbers, for overriding the values of parameters in the Database.
     callables : dict, optional
@@ -204,7 +204,7 @@ def build_phase_records(dbf, comps, phases, conds, models, output='GM',
         'internal_cons_hess': {},
     }
     phase_records = {}
-    state_variables = sorted(get_state_variables(models=models, conds=conds), key=str)
+    state_variables = sorted(get_state_variables(models=models, conds=state_variables), key=str)
     param_symbols, param_values = extract_parameters(parameters)
 
     if callables.get(output) is None: