--- a/pycalphad/tests/test_equilibrium.py
+++ b/pycalphad/tests/test_equilibrium.py
@@ -63,13 +63,23 @@ def test_phase_records_passed_to_equilibrium():
     models = instantiate_models(ALFE_DBF, comps, my_phases)
     phase_records = build_phase_records(ALFE_DBF, comps, my_phases, conds, models)
 
-    # Without models passed
-    eqx = equilibrium(ALFE_DBF, comps, my_phases, conds, verbose=True, phase_records=phase_records)
-    assert_allclose(eqx.GM.values.flat[0], -9.608807e4)
-
     # With models passed
     eqx = equilibrium(ALFE_DBF, comps, my_phases, conds, verbose=True, model=models, phase_records=phase_records)
     assert_allclose(eqx.GM.values.flat[0], -9.608807e4)
+
+
+def test_missing_models_with_phase_records_passed_to_equilibrium_raises():
+    "equilibrium should raise an error if all the active phases are not included in the phase_records"
+    my_phases = ['LIQUID', 'FCC_A1', 'HCP_A3', 'AL5FE2', 'AL2FE', 'AL13FE4', 'AL5FE4']
+    comps = ['AL', 'FE', 'VA']
+    conds = {v.T: 1400, v.P: 101325, v.N: 1.0, v.X('AL'): 0.55}
+
+    models = instantiate_models(ALFE_DBF, comps, my_phases)
+    phase_records = build_phase_records(ALFE_DBF, comps, my_phases, conds, models)
+
+    with pytest.raises(ValueError):
+        # model=models NOT passed
+        equilibrium(ALFE_DBF, comps, my_phases, conds, verbose=True, phase_records=phase_records)
 
 
 def test_missing_phase_records_passed_to_equilibrium_raises():
@@ -79,11 +89,19 @@ def test_missing_phase_records_passed_to_equilibrium_raises():
     comps = ['AL', 'FE', 'VA']
     conds = {v.T: 1400, v.P: 101325, v.N: 1.0, v.X('AL'): 0.55}
 
-    models = instantiate_models(ALFE_DBF, comps, subset_phases)
-    phase_records = build_phase_records(ALFE_DBF, comps, subset_phases, conds, models)
+    models = instantiate_models(ALFE_DBF, comps, my_phases)
+    phase_records = build_phase_records(ALFE_DBF, comps, my_phases, conds, models)
 
+    models_subset = instantiate_models(ALFE_DBF, comps, subset_phases)
+    phase_records_subset = build_phase_records(ALFE_DBF, comps, subset_phases, conds, models_subset)
+
+    # Under-specified models
     with pytest.raises(ValueError):
-        equilibrium(ALFE_DBF, comps, my_phases, conds, verbose=True, phase_records=phase_records)
+        equilibrium(ALFE_DBF, comps, my_phases, conds, verbose=True, model=models_subset, phase_records=phase_records)
+
+    # Under-specified phase_records
+    with pytest.raises(ValueError):
+        equilibrium(ALFE_DBF, comps, my_phases, conds, verbose=True, model=models, phase_records=phase_records_subset)
 
 
 @pytest.mark.solver