Changes:

- `build_phase_records`
  - Defaults to build derivatives of the output. This is the more common case and SymEngine makes the compilation fast enough that's it's not as much of a burden to build them by default.
  - The `comps` argument now accepts string components or `Species` objects via `unpack_components`. Before this change, only `Species` objects were allowed.
- `calculate` and `equilibrium`
  - Accept the `phase_records` keyword argument 
  - The `callables` keyword argument is used if the phase records are constructed. If the `phase_records` argument is supplied by the caller, the `callables` argument is ignored.
  - `equilibrium` passes `phase_records` down to `calculate` instead of rebuilding them in `calculate`
  - Both functions check that all the active phases have entries in `phase_records` and raise an error if not. We currently don't check that the `output` is correct at all
- This approach is relatively simple, but one gap is that there's no mechanism to pass `phase_records` in to `equilibrium` that get re-used by `_eqcalculate`. `calculate` will always build `phase_records` when called from `_eqcalculate` (the `callables` can still be passed down for now).