@bocklund 2021-07-19
Calling `equilibrium` in tight loops could be really expensive because the time to continually rebuild the phase records. After making the changes in this PR, I was able to call `build_phase_records` in the scheil package and pass them to equilibrium and see huge boost in performance, see these profiles of a single simulation. The benchmark images below show a reduction from 491 seconds -> 8 seconds by eliminating the redundant calls to `build_phase_records` every time `equilibrium` is called:


## Before: passing the callables from `build_callables` to `equilibrium` (491 seconds, 95% of time in `build_phase_records` called 1204 times)

![image](https://user-images.githubusercontent.com/7681751/126206947-3dfa68ce-d449-425e-9209-64aeaadbd1e4.png)

## After: passing the phase records from `build_phase_records` to `equilibrium` (8.6 seconds, 0.9 seconds of time in `build_phase_records` called one time)

![image](https://user-images.githubusercontent.com/7681751/126206958-60ddb989-b96a-4bd9-adb4-2b8d041c4df8.png)


## Why didn't we do this before?

Before the new minimizer was implemented, the choice of external conditions was tightly coupled with `PhaseRecord` objects via `Model.get_multiphase_constraints`. That is, changing the conditions for an equilibrium calculation in an A-B binary system from `N`, `P`, `T`, `X(B)` to `N`, `P`, `T`, `X(A)` required a new set of phase records. This tight coupling was what led to passing around `callables`, which only depended on the state variables and not the external conditions. Since the conditions in the new minimizer are implemented inside the minimizer itself, each `PhaseRecord` now only needs to know about the state variables and the internal constraints for that phase.